app/requirements.txt
asgiref==3.10.0
Django==5.2.8
django-allauth==65.13.1
django-stubs==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.15.1
drf-spectacular==0.29.0
gunicorn==23.0.0
itemadapter==0.12.2
pre_commit==4.5.0
psycopg2-binary==2.9.11
pyright==1.1.407
python-decouple==3.8
scrapy==2.13.3
selenium==4.38.0
supervisor==4.3.0
